# backend/src/users/dto/user.dto.ts
import { Exclude } from 'class-transformer';
import { ApiProperty } from '@nestjs/swagger';

export class UserDto {
  @ApiProperty()
  id: string;

  @ApiProperty()
  mobileNumber: string;

  @ApiProperty()
  firstName: string;

  @ApiProperty()
  lastName: string;

  @ApiProperty()
  avatarUrl: string;

  @ApiProperty()
  profileData: object;

  @ApiProperty()
  createdAt: Date;

  @ApiProperty()
  updatedAt: Date;

  @Exclude()
  otp: string;

  @Exclude()
  otpExpiresAt: Date;

  constructor(partial: Partial<UserDto>) {
    Object.assign(this, partial);
  }
}