# sdk/swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/auth/request-otp":{"post":{"operationId":"AuthController_requestOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestOtpDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/verify-otp":{"post":{"operationId":"AuthController_verifyOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyOtpDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"tags":["Auth"]}},"/users/me":{"get":{"operationId":"UsersController_getProfile","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"security":[{"bearer":[]}],"tags":["Users"]},"patch":{"operationId":"UsersController_updateProfile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"security":[{"bearer":[]}],"tags":["Users"]}},"/conversations":{"post":{"operationId":"ConversationsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateConversationDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConversationDto"}}}}},"security":[{"bearer":[]}],"tags":["Conversations"]},"get":{"operationId":"ConversationsController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ConversationDto"}}}}}},"security":[{"bearer":[]}],"tags":["Conversations"]}},"/conversations/{id}":{"get":{"operationId":"ConversationsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConversationDto"}}}}},"security":[{"bearer":[]}],"tags":["Conversations"]}},"/conversations/{id}/participants":{"post":{"operationId":"ConversationsController_addParticipants","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddParticipantsDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConversationDto"}}}}},"security":[{"bearer":[]}],"tags":["Conversations"]}},"/conversations/{id}/messages":{"get":{"operationId":"MessagesController_findAll","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"before","required":false,"in":"query","description":"Cursor for pagination (a message ID)","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"The maximum number of messages to return","schema":{"minimum":1,"maximum":100,"default":20,"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessageDto"}}}}}},"security":[{"bearer":[]}],"tags":["Messages"]}},"/conversations/{id}/tasks":{"get":{"operationId":"TasksController_findAll","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaskDto"}}}}}},"security":[{"bearer":[]}],"tags":["Tasks"]}}},"info":{"title":"Simflo API","description":"The Simflo API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RequestOtpDto":{"type":"object","properties":{"mobileNumber":{"type":"string","description":"The mobile number to send the OTP to.","example":"+15555555555"}},"required":["mobileNumber"]},"VerifyOtpDto":{"type":"object","properties":{"mobileNumber":{"type":"string","description":"The mobile number to verify.","example":"+15555555555"},"otp":{"type":"string","description":"The 4-digit OTP.","example":"1234"}},"required":["mobileNumber","otp"]},"UserDto":{"type":"object","properties":{"id":{"type":"string"},"mobileNumber":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"avatarUrl":{"type":"string"},"profileData":{"type":"object"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","mobileNumber","firstName","lastName","avatarUrl","profileData","createdAt","updatedAt"]},"AuthResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"},"user":{"$ref":"#/components/schemas/UserDto"}},"required":["accessToken","user"]},"UpdateProfileDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"avatarUrl":{"type":"string"},"profileData":{"type":"object"}}},"ParticipantDto":{"type":"object","properties":{"userId":{"type":"string"},"role":{"type":"string","enum":["ADMIN","MEMBER","SUBSCRIBER","NOTIFIER"]}},"required":["userId","role"]},"CreateConversationDto":{"type":"object","properties":{"title":{"type":"string"},"type":{"type":"string","enum":["DIRECT_MESSAGE","GROUP_CHAT"]},"tags":{"type":"array","items":{"type":"string"}},"participants":{"type":"array","items":{"$ref":"#/components/schemas/ParticipantDto"}}},"required":["title","type","participants"]},"ConversationParticipantDto":{"type":"object","properties":{"id":{"type":"string"},"role":{"type":"string","enum":["ADMIN","MEMBER","SUBSCRIBER","NOTIFIER"]},"user":{"$ref":"#/components/schemas/UserDto"}},"required":["id","role","user"]},"ConversationDto":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"type":{"type":"string","enum":["DIRECT_MESSAGE","GROUP_CHAT"]},"tags":{"type":"array","items":{"type":"string"}},"participants":{"type":"array","items":{"$ref":"#/components/schemas/ConversationParticipantDto"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","title","type","tags","participants","createdAt","updatedAt"]},"AddParticipantsDto":{"type":"object","properties":{"participants":{"type":"array","items":{"$ref":"#/components/schemas/ParticipantDto"}}},"required":["participants"]},"MessageContentDto":{"type":"object","properties":{"text":{"type":"string"}},"required":["text"]},"MessageDto":{"type":"object","properties":{"id":{"type":"string"},"sender":{"$ref":"#/components/schemas/UserDto"},"conversationId":{"type":"string"},"content":{"$ref":"#/components/schemas/MessageContentDto"},"status":{"type":"string","enum":["SENT","DELIVERED","READ"]},"createdAt":{"format":"date-time","type":"string"}},"required":["id","sender","conversationId","content","status","createdAt"]},"TaskDto":{"type":"object","properties":{"id":{"type":"string"},"message":{"$ref":"#/components/schemas/MessageDto"},"assignee":{"$ref":"#/components/schemas/UserDto"},"status":{"type":"string","enum":["TODO","IN_PROGRESS","DONE"]},"dueDate":{"format":"date-time","type":"string"}},"required":["id","message","status"]}}}}